- Renamed Error to ErrorComponent to prevent clashes.
